+++ outputs.tf
@@ -1,13 +1,13 @@
 # ---------------------------------------------------------------------------------------------------
-# version  1.3
+# version  1.5
 # Library: https://github.com/Frankie116/my-library.git
 # outputs
 # ---------------------------------------------------------------------------------------------------
 
 
-output "my-alb-hostname" {
-  value = aws_lb.my-alb.dns_name
-}
+# output "my-alb-hostname" {
+#   value = aws_lb.my-alb.dns_name
+# }
 
 # 6a-route53.tf ----------------------------------------------------------
 output this-new-dns-name {
@@ -23,8 +23,7 @@ output my-docker-port {
 }
 
 # 09c-template-file ----------------------------------------------------------
-output my-docker-image {
-  description          = "Name of the docker image"
-  value                = var.my-docker-image
+output my-ecs-cd-template {
+  description          = "Name of the container definition"
+  value                = var.my-ecs-cd-template 
 }
-